ContentViewModel unified MainViewModel
Delete ContentViewModel
